mt_chess_piece_get_index():Use correct type for ID.

+++ mt_chess_piece.c
@@ -17,7 +17,8 @@
 #include "mt_chess_color.h"
 #include "mt_chess_type.h"
 
-int mt_chess_piece_get_index(struct mt_chess_piece * const pieces, int const id)
+int mt_chess_piece_get_index(
+        struct mt_chess_piece * const pieces, uint8_t const id)
 {
     assert(pieces != NULL);
     // Also assuming that array the pointer points to has 2 x 2 x 8 items.
